lift and shift analytics component from fee transfer. Refactor to accept block input
